godfather/requirements.txt
pycrypto
requests
numpy
openalpr